+++ duck/http/session/engine.py
@@ -11,8 +11,12 @@ from functools import wraps
 
 from duck.settings import SETTINGS
 from duck.utils.importer import import_module_once
+from duck.utils.asyncio import raise_if_in_async_context
+from duck.contrib.sync import ensure_async
 from duck.logging import logger
 
+
+# TODO: Need to make the SESSION API fully async-compatible
 
 session_connector_mod = import_module_once("duck.http.session.session_storage_connector")
 
@@ -67,6 +71,13 @@ class SessionStore(dict):
         """
         return self._modified
     
+    @property
+    def loaded(self) -> bool:
+        """
+        Returns the state whether the session has been loaded.
+        """
+        return self._loaded
+    
     @modified.setter
     def modified(self, what: bool):
         """
@@ -78,14 +89,10 @@ class SessionStore(dict):
         """
         Returns whether the session data is worthy to be saved, this is the lazy behavior of Duck.
         """
-        if (self and self.session_key 
-            and self._modified
-            and not ("expiry_date" in self
-            and len(self) == 1)
-        ):
-            # The session data is not empty, has been modified, session data length is (not 1 and containing only the key `expiry_date`)
-            return True
-        return False
+        if not self.loaded:
+            # If session hasn't been loaded, we consider it as not modified to avoid unnecessary saves
+            return False
+        return self.modified
 
     @staticmethod
     def generate_session_id() -> str:
@@ -93,6 +100,14 @@ class SessionStore(dict):
         Generates and returns a random session ID.
         """
         return str(uuid.uuid4())
+    
+    def session_expired(self) -> bool:
+        """
+        Returns boolean on whether if the session has expired depending on expiry set on session.
+        """
+        expiry_date = self.get_expiry_date()
+        now = datetime.datetime.utcnow()
+        return now >= expiry_date
 
     def get_expiry_age(self):
         """
@@ -121,9 +136,9 @@ class SessionStore(dict):
         Sets the session expiry.
 
         Args:
-                expiry Optional[Union[int, float, datetime.datetime, datetime.timedelta]]: Float or int represents the seconds to expire from now and None represents the now plus the default session max_age.
+            expiry Optional[Union[int, float, datetime.datetime, datetime.timedelta]]: Float or int represents the seconds to expire from now and None represents the now plus the default session max_age.
         """
-        if not expiry:
+        if expiry is None:
             self["expiry_date"] = (
                 datetime.datetime.utcnow()
                 + datetime.timedelta(seconds=self.get_expiry_age())
@@ -132,7 +147,7 @@ class SessionStore(dict):
         elif isinstance(expiry, (datetime.datetime, datetime.timedelta)):
             self["expiry_date"] = expiry
 
-        elif isinstance(expiry, int, float):
+        elif isinstance(expiry, (int, float)):
             self["expiry_date"] = (
                 datetime.datetime.utcnow()
                 + datetime.timedelta(seconds=expiry)
@@ -147,7 +162,7 @@ class SessionStore(dict):
     def check_session_storage_connector(method):
         """
         Decorator to ensure a valid session storage connector is present.
-    
+
         Validates:
             - Attribute exists
             - Attribute is not None
@@ -173,52 +188,83 @@ class SessionStore(dict):
     
         # Return wrapper
         return wrapper
+    
+    @staticmethod
+    def ensure_session_loaded(method):
+        """
+        Decorator which ensures that the session is loaded.
+        """
+        @wraps(method)
+        def wrapper(self, *args, **kwargs):
+            if not self.loaded:
+                self.load()
+            
+            # Execute the decorated method
+            return method(self, *args, **kwargs)
+    
+        # Return wrapper
+        return wrapper
 
     @check_session_storage_connector
     def load(self) -> dict:
         """
         Loads the session from storage.
         """
-        if self._loaded and not self.disable_warnings:
+        if self.loaded and not self.disable_warnings:
             logger.warn(
                 f"{self.__class__.__name__} is already loaded; reloading may be inefficient."
             )
         
-        # Update session data
+        # Retrieve session data from storage, if session key is invalid or session doesn't exist, get_session should return None, so we can safely fallback to empty dict
         session_data = self.session_storage_connector.get_session(self.session_key) or {}
-        self.update(session_data)
         
-        if not self._loaded:
+        # Update session store with retrieved data, this will also set the modified flag if data is not empty
+        super().update(session_data) # Avoids recursion error
+        
+        if not self.loaded:
             self._modified = False # If session hasn't been loaded for the first time, set _modified to False
-        
+            
         # Update state and return session data
         self._loaded = True
         return session_data
 
+    def save(self):
+        """
+        Save the session
+        """
+        raise_if_in_async_context("Please use 'async_save' method instead.")
+        self._save()
+
     @check_session_storage_connector
-    def save(self, *_):
+    @ensure_session_loaded
+    def _save(self):
         """
         Saves the session to storage.
         """
         if not self.session_key:
             raise ValueError("Session key is not set or invalid.")
         
+        # Normalize session data
         session_data = dict(self)
-        expiry_date = self.get_expiry_date()
-        now = datetime.datetime.utcnow()
-
-        if now < expiry_date:
+        
+        if not self.session_expired():
             # Session is not expired
-            expiry_age = expiry_date - now
-            expiry_age = expiry_age.total_seconds()
+            expiry_age = self.get_expiry_age()
+            
+            # Set the session in the real storage
             self.session_storage_connector.set_session(
                 self.session_key,
                 session_data,
                 expiry=expiry_age,
             )
+
+            # Save the current state of session storage
             self.session_storage_connector.save()
+        
         else:
             raise SessionExpired("Cannot save an expired session, use `set_expiry` to reset the session expiry.")
+        
+        # Reset session modification
         self._modified = False  # reset session modification.
 
     @check_session_storage_connector
@@ -230,18 +276,28 @@ class SessionStore(dict):
             session_key (Optional[str]): The session key or None if you want to use the current session key.
         """
         session_key = session_key or self.session_key
+        
         try:
             return bool(self.session_storage_connector.get_session(session_key))
         except KeyError:
             return False
 
     @check_session_storage_connector
-    def create(self):
+    def assign_new_session_key(self) -> str:
         """
         Creates a new session with a new session key.
         """
         self.session_key = self.generate_session_id()
+        return self.session_key
         
+    @check_session_storage_connector
+    @ensure_session_loaded
+    def get(self, *args, **kw):
+        """
+        Return value for a key.
+        """
+        return super().get(*args, **kw)
+
     @check_session_storage_connector
     def delete(self, session_key: Optional[str] = None):
         """
@@ -254,49 +310,78 @@ class SessionStore(dict):
         self.session_storage_connector.delete_session(session_key)
         self.clear()
 
+    @ensure_session_loaded
     def update(self, data: dict):
         """
         Overrides the update method to ensure items are tracked for modification.
         """
         super().update(data)
+        
         if data:
             self._modified = True
 
+    @ensure_session_loaded
     def clear(self):
         """
         Clears all session data.
         """
         is_empty = bool(self or None)
+        
         super().clear()
+        
         if not is_empty:
             self._modified = True
 
+    @ensure_session_loaded
     def pop(self, *args, **kwargs):
         """
         Pops some session data.
         """
         data = super().pop(*args, **kwargs)
-        self._modified = True
+        
+        # If data is not empty, we consider the session modified, otherwise we don't to avoid unnecessary saves
+        if data:
+            self._modified = True
+        
+        # Return popped data
         return data
 
+    @ensure_session_loaded
     def popitem(self, *args, **kwargs):
         """
         Pops some session data.
         """
         data = super().popitem(*args, **kwargs)
-        self._modified = True
-        return data
 
+        # If data is not empty, we consider the session modified, otherwise we don't to avoid unnecessary saves
+        if data:
+            self._modified = True
+        
+        # Return popped data
+        return data
+    
+    # SOME ASYNC API's
+    async def async_save(self):
+        """
+        Asynchronously save session.
+        """
+        await ensure_async(self._save)()
+
+    @ensure_session_loaded
     def __setitem__(self, key, value):
         super().__setitem__(key, value)
         self._modified = True
 
+    @ensure_session_loaded
     def __getitem__(self, key):
         value = super().__getitem__(key)
+        return value
 
+    @ensure_session_loaded
     def __delitem__(self, key):
         super().__delitem__(key)
         self._modified = True
 
+    @ensure_session_loaded
     def __repr__(self):
         return f"<{self.__class__.__name__} {dict(self)}>"